Add .bash_profile
Recreated the missing .bash_profile which was removed in a previous commit by mistake and moved a lot of stuff between the different bash startup files.
This commit FIXES a persistent problem which occured when using sshfs to load the cluster filesystem on the local machine. Reason was that .bashrc produced output which killed the sshfs process.